LRQA-35482 Consolidate BasePoshiComment with PoshiComment. Consolidate BasePoshiElement with PoshiElement Tighten up PoshiNode generics
