+++ poshi-runner/src/main/java/com/liferay/poshi/runner/elements/PoshiNode.java
@@ -19,11 +19,12 @@ import org.dom4j.Node;
 /**
 * @author Michael Hashimoto
  */
-public interface PoshiNode<A extends B, B> extends Node {
+public interface PoshiNode<A extends Node, B extends PoshiNode<A, B>>
+	extends Node {
 
-	public A clone(B node);
+	public B clone(A node);
 
-	public A clone(String readableSyntax);
+	public B clone(String readableSyntax);
 
 	public void parseReadableSyntax(String readableSyntax);
